requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kafka==1.3.5
kafka-python==2.0.2
MarkupSafe==2.1.2
minio==7.1.14
psutil==5.9.4
public-ip==0.12
pydantic==1.10.7
pymongo==4.3.3
pyrsistent==0.19.3
python-on-whales==0.60.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
